+++ botv2/build.rs
@@ -14,6 +14,13 @@ pub fn modules() -> Vec<crate::silverpelt::Module> {
         {module_func_list}
     ]
 }
+
+/// Module id list
+pub fn module_ids() -> Vec<&'static str> {
+    vec![
+        {module_ids_list}
+    ]
+}
     "#;
 
     let mut module_list = Vec::new();
@@ -75,7 +82,7 @@ pub fn modules() -> Vec<crate::silverpelt::Module> {
     let module_list_final = MODULE_TEMPLATE
         .replace("{module_use_list}", &module_use_list)
         .replace("{module_func_list}", &module_func_list)
-        .replace("{module_ids_list}", &module_ids_list.join("\n        ")); // Not used currently but may be used in the future
+        .replace("{module_ids_list}", &module_ids_list.join("\n        "));
 
     let module_list_final = module_list_final.trim().to_string();